restricted pages

+++ base/views.py
@@ -1,4 +1,5 @@
 
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.db.models import Q
@@ -6,6 +7,7 @@ from .models import Room, Topic
 from .forms import RoomForm
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
 
 
 # Create your views here.
@@ -31,6 +33,7 @@ def room(request, pk):
     return render(request, 'pages/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -42,9 +45,13 @@ def createRoom(request):
     return render(request, 'pages/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -55,7 +62,12 @@ def updateRoom(request, pk):
     return render(request, 'pages/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!')
+
     room = Room.objects.get(id=pk)
     if request.method == 'POST':
         room.delete()
@@ -64,6 +76,8 @@ def deleteRoom(request, pk):
 
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
 
     if request.method == 'POST':
         username = request.POST.get('username')
